Moved examples from cluster to child_process Closes HowProgrammingWorks/InterProcessCommunication#3
