+++ workers/worker.js
@@ -1,22 +1,18 @@
 'use strict';
 
-module.exports = () => {
+console.log(process.pid + '\tHELLO');
 
-  console.log('Hello from worker ' + process.pid + ' ' + api.cluster.worker.id);
+const caltulations = item => item * 2;
 
-  const caltulations = item => item * 2;
+process.on('message', (message) => {
 
-  process.on('message', (message) => {
+  console.log(
+    process.pid + '\t' +
+    'FROM master: ' + JSON.stringify(message)
+  );
 
-    console.log(
-      'message to worker ' + process.pid +
-      ' from master: ' + JSON.stringify(message)
-    );
-
-    process.send({
-      result: message.task.map(caltulations)
-    });
-
+  process.send({
+    result: message.task.map(caltulations)
   });
 
-};
+});
